requirements.txt
environs==9.*
python-telegram-bot==13.*
requests==2.28.*
